requirements.txt
annotated-types==0.7.0
apparmor==4.1.2
argcomplete==3.6.2
asgiref==3.9.1
attrs==25.3.0
bearlibterminal==0.15.8
blessed==1.21.0
Brlapi==0.8.6
certifi==2025.8.3
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
copyparty==1.19.7
cryptography==45.0.7
dbus-python==1.4.0
defusedxml==0.7.1
distro==1.9.0
Django==5.2.6
docutils==0.21.2
filelock==3.19.1
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
legacy-cgi==2.6.3
LibAppArmor==4.1.2
libtorrent==2.0.11
llvmlite==0.44.0
louis==3.34.0
Markdown==3.9
Markups==4.1.1
MarkupSafe==3.0.2
netsnmp-python==1.0a1
nftables==0.1
noise==1.2.2
numba==0.61.2
numpy==2.2.6
packaging==25.0
panda==0.3.1
pandas==2.3.1
pillow==11.3.0
pipx==1.7.1
platformdirs==4.4.0
pluggy==1.6.0
psutil==7.0.0
pwquality==1.4.5
pyaml==25.7.0
pycairo==1.28.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pygame==2.6.1
pygdbmi==0.11.0.0
Pygments==2.19.2
PyGObject==3.52.3
PyQt6==6.9.1
PyQt6_sip==13.10.2
pytest==8.4.1
python-dateutil==2.9.0.post0
python-markdown-math==0.9
python-statemachine==2.5.0
python-xlib==0.33
pytz==2025.2
pyxdg==0.28
PyYAML==6.0.2
requests==2.32.4
scipy==1.16.1
sentry-sdk==2.37.1
setproctitle==1.3.7
setuptools==80.9.0
six==1.17.0
smbus==1.1
sqlparse==0.5.3
statemachine==0.1
TBB==0.2
tcod==19.2.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ufw==0.36.2
urllib3==2.5.0
userpath==1.9.2
wcwidth==0.2.13
wheel==0.45.1
wxPython==4.2.3
